scraper/requirements.txt
selenium
python-dotenv
sqlalchemy>=2.0.0
psycopg2-binary
alembic